+++ app/project/[projectId]/_components/ProjectSection/ProjectHeader.tsx
@@ -1,20 +1,18 @@
 "use client";
 import React from "react";
-import { useMutation, UseQueryResult, useQueryClient } from "@tanstack/react-query";
+import { UseQueryResult } from "@tanstack/react-query";
 import { useRouter } from "next/navigation";
 import WishButtonAndCount from "@/app/_components/Button/WishButton";
 import { ProjectResponse } from "@/app/_apis/schema/projectResponse";
-import { projectQueryKey } from "@/app/_queryFactory/projectQuery";
 import { createDate } from "@/app/_utils/createDate";
 import { JOB_CATEGORIES_KR } from "@/app/_constants/JobCategoryData";
-import { projectApi } from "@/app/_apis/projectApi";
-import { useToast } from "@/app/_context/ToastContext";
 import useModal from "@/app/_hooks/useModal";
 import WarningModal from "@/app/_components/Modal/WarningModal";
 import { revalidateTagAction } from "@/app/_utils/revalidationAction";
 import { WINDOW_BOUNDARY } from "@/app/_constants/WindowSize";
 import useBrowserSize from "@/app/_hooks/useBrowserSize";
 import { useProjectDetail } from "@/app/_hooks/reactQuery/useProjectQuery";
+import useProjectMutation from "@/app/_hooks/mutations/useProjectMutation";
 import MenuDropBox from "../Project/DropBox/MenuDropBox";
 import SocialDropBox from "../Project/DropBox/SocialDropBox";
 
@@ -25,30 +23,14 @@ interface Props {
 type JobCategory = keyof typeof JOB_CATEGORIES_KR;
 
 function ProjectHeader({ projectId }: Props) {
-  const { addToast } = useToast();
   const { openModal: deleteModal, handleModalOpen: openDeleteModal, handleModalClose: closeDeleteModal } = useModal();
-  const queryClient = useQueryClient();
   const router = useRouter();
 
   const { windowWidth } = useBrowserSize();
   const { MB } = WINDOW_BOUNDARY.MAX;
   const { TBC } = WINDOW_BOUNDARY.MIN;
 
-  const mutation = useMutation({
-    mutationFn: () => {
-      return projectApi.deleteProject(projectId);
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: projectQueryKey.list().queryKey,
-      });
-      addToast("프로젝트가 삭제되었습니다", "success");
-    },
-    onError: error => {
-      console.error("Error:", error);
-      addToast("프로젝트 삭제 오류가 발생했습니다", "error");
-    },
-  });
+  const { projectDeleteMutation } = useProjectMutation(projectId);
 
   const { data: project }: UseQueryResult<ProjectResponse, Error> = useProjectDetail(projectId);
 
@@ -60,7 +42,7 @@ function ProjectHeader({ projectId }: Props) {
 
   const handleDeleteProject = () => {
     revalidateTagAction("projectList");
-    mutation.mutate();
+    projectDeleteMutation.mutate();
     router.push("/main");
     closeDeleteModal();
   };
